Refactored MainActivity, put openGl in own class To make MainActivity more readable, seperated activity specific code and openGl surface rendering. Made a new CameraEyeTransformer class that contain everything related to this.
