+++ app/src/main/java/no/ntnu/tpg4850/frugalar/CameraEyeTransformer.java
@@ -1,0 +1,189 @@
+package no.ntnu.tpg4850.frugalar;
+
+import android.opengl.GLES20;
+import android.util.Log;
+
+import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
+import java.nio.FloatBuffer;
+import java.nio.ShortBuffer;
+
+import javax.microedition.khronos.opengles.GL10;
+
+/**
+ * Created by Olav on 23.02.2015.
+ */
+public class CameraEyeTransformer {
+
+    private static final String TAG = "CameraEyeTransformer";
+    private static final int GL_TEXTURE_EXTERNAL_OES = 0x8D65;
+    // number of coordinates per vertex in this array
+    static final int COORDS_PER_VERTEX = 2;
+    private final int vertexStride = COORDS_PER_VERTEX * 4; // 4 bytes per vertex
+
+
+    private FloatBuffer vertexBuffer, textureVerticesBuffer;
+    private ShortBuffer drawListBuffer;
+
+    private short drawOrder[] =  {0, 2, 1, 1, 2, 3 }; // order to draw vertices
+    private short drawOrder2[] = {2, 0, 3, 3, 0, 1}; // order to draw vertices
+
+    private int mProgram;
+    private int mPositionHandle;
+    private int mColorHandle;
+    private int mTextureCoordHandle;
+
+    private final String vertexShaderCode =
+            "attribute vec4 position;" +
+                    "attribute vec2 inputTextureCoordinate;" +
+                    "varying vec2 textureCoordinate;" +
+                    "void main()" +
+                    "{"+
+                    "gl_Position = position;"+
+                    "textureCoordinate = inputTextureCoordinate;" +
+                    "}";
+
+    private final String fragmentShaderCode =
+            "#extension GL_OES_EGL_image_external : require\n"+
+                    "precision mediump float;" +
+                    "varying vec2 textureCoordinate;\n" +
+                    "uniform samplerExternalOES s_texture;\n" +
+                    "void main(void) {" +
+                    "  gl_FragColor = texture2D( s_texture, textureCoordinate );\n" +
+                    "}";
+
+    static float textureVertices[] = {
+            0.0f, 1.0f,  // A. left-bottom
+            1.0f, 1.0f,  // B. right-bottom
+            0.0f, 0.0f,  // C. left-top
+            1.0f, 0.0f   // D. right-top
+    };
+
+    static float squareVertices[] = { // in counterclockwise order:
+            -1.0f, -1.0f,   // 0.left - mid
+            1.0f, -1.0f,   // 1. right - mid
+            -1.0f, 1.0f,   // 2. left - top
+            1.0f, 1.0f,   // 3. right - top
+    };
+
+    public CameraEyeTransformer() {
+
+    }
+
+    public void createSurface() {
+        GLES20.glClearColor(0.1f, 0.1f, 0.1f, 0.5f); // Dark background so text shows up well
+
+        ByteBuffer bb = ByteBuffer.allocateDirect(squareVertices.length * 4);
+        bb.order(ByteOrder.nativeOrder());
+        vertexBuffer = bb.asFloatBuffer();
+        vertexBuffer.put(squareVertices);
+        vertexBuffer.position(0);
+
+
+        ByteBuffer dlb = ByteBuffer.allocateDirect(drawOrder.length * 2);
+        dlb.order(ByteOrder.nativeOrder());
+        drawListBuffer = dlb.asShortBuffer();
+        drawListBuffer.put(drawOrder);
+        drawListBuffer.position(0);
+
+
+        ByteBuffer bb2 = ByteBuffer.allocateDirect(textureVertices.length * 4);
+        bb2.order(ByteOrder.nativeOrder());
+        textureVerticesBuffer = bb2.asFloatBuffer();
+        textureVerticesBuffer.put(textureVertices);
+        textureVerticesBuffer.position(0);
+
+        int vertexShader = loadGLShader(GLES20.GL_VERTEX_SHADER, vertexShaderCode);
+        int fragmentShader = loadGLShader(GLES20.GL_FRAGMENT_SHADER, fragmentShaderCode);
+
+        mProgram = GLES20.glCreateProgram();             // create empty OpenGL ES Program
+        GLES20.glAttachShader(mProgram, vertexShader);   // add the vertex shader to program
+        GLES20.glAttachShader(mProgram, fragmentShader); // add the fragment shader to program
+        GLES20.glLinkProgram(mProgram);
+    }
+
+    public void drawEye(int texture) {
+        this.clearGL();
+
+        GLES20.glUseProgram(mProgram);
+
+        GLES20.glActiveTexture(GL_TEXTURE_EXTERNAL_OES);
+        GLES20.glBindTexture(GL_TEXTURE_EXTERNAL_OES, texture);
+
+        mPositionHandle = GLES20.glGetAttribLocation(mProgram, "position");
+        GLES20.glEnableVertexAttribArray(mPositionHandle);
+        GLES20.glVertexAttribPointer(mPositionHandle, COORDS_PER_VERTEX, GLES20.GL_FLOAT,
+                false,vertexStride, vertexBuffer);
+
+        mTextureCoordHandle = GLES20.glGetAttribLocation(mProgram, "inputTextureCoordinate");
+        GLES20.glEnableVertexAttribArray(mTextureCoordHandle);
+        GLES20.glVertexAttribPointer(mTextureCoordHandle, COORDS_PER_VERTEX, GLES20.GL_FLOAT,
+                false,vertexStride, textureVerticesBuffer);
+
+        mColorHandle = GLES20.glGetAttribLocation(mProgram, "s_texture");
+
+        GLES20.glDrawElements(GLES20.GL_TRIANGLES, drawOrder.length,
+                GLES20.GL_UNSIGNED_SHORT, drawListBuffer);
+
+
+        // Disable vertex array
+        GLES20.glDisableVertexAttribArray(mPositionHandle);
+        GLES20.glDisableVertexAttribArray(mTextureCoordHandle);
+    }
+
+    public static int createTexture() {
+        int[] texture = new int[1];
+
+        GLES20.glGenTextures(1, texture, 0);
+        GLES20.glBindTexture(GL_TEXTURE_EXTERNAL_OES, texture[0]);
+        GLES20.glTexParameterf(GL_TEXTURE_EXTERNAL_OES,
+                GL10.GL_TEXTURE_MIN_FILTER,GL10.GL_LINEAR);
+        GLES20.glTexParameterf(GL_TEXTURE_EXTERNAL_OES,
+                GL10.GL_TEXTURE_MAG_FILTER, GL10.GL_LINEAR);
+        GLES20.glTexParameteri(GL_TEXTURE_EXTERNAL_OES,
+                GL10.GL_TEXTURE_WRAP_S, GL10.GL_CLAMP_TO_EDGE);
+        GLES20.glTexParameteri(GL_TEXTURE_EXTERNAL_OES,
+                GL10.GL_TEXTURE_WRAP_T, GL10.GL_CLAMP_TO_EDGE);
+
+        return texture[0];
+    }
+
+    public void clearGL() {
+        GLES20.glClear(GLES20.GL_COLOR_BUFFER_BIT | GLES20.GL_DEPTH_BUFFER_BIT);
+    }
+
+    private int loadGLShader(int type, String code) {
+        int shader = GLES20.glCreateShader(type);
+        GLES20.glShaderSource(shader, code);
+        GLES20.glCompileShader(shader);
+
+        // Get the compilation status.
+        final int[] compileStatus = new int[1];
+        GLES20.glGetShaderiv(shader, GLES20.GL_COMPILE_STATUS, compileStatus, 0);
+
+        // If the compilation failed, delete the shader.
+        if (compileStatus[0] == 0) {
+            Log.e(TAG, "Error compiling shader: " + GLES20.glGetShaderInfoLog(shader));
+            GLES20.glDeleteShader(shader);
+            shader = 0;
+        }
+
+        if (shader == 0) {
+            throw new RuntimeException("Error creating shader.");
+        }
+
+        return shader;
+    }
+
+    /**
+     * Checks if we've had an error inside of OpenGL ES, and if so what that error is.
+     * @param func
+     */
+    private static void checkGLError(String func) {
+        int error;
+        while ((error = GLES20.glGetError()) != GLES20.GL_NO_ERROR) {
+            Log.e(TAG, func + ": glError " + error);
+            throw new RuntimeException(func + ": glError " + error);
+        }
+    }
+}
