+++ app/src/main/java/no/ntnu/tpg4850/frugalar/MainActivity.java
@@ -40,56 +40,8 @@ import java.nio.ShortBuffer;
 public class MainActivity extends CardboardActivity implements CardboardView.StereoRenderer, OnFrameAvailableListener {
 
     private static final String TAG = "MainActivity";
-    private static final int GL_TEXTURE_EXTERNAL_OES = 0x8D65;
+
     private Camera myCamera = null;
-
-    private final String vertexShaderCode =
-            "attribute vec4 position;" +
-                    "attribute vec2 inputTextureCoordinate;" +
-                    "varying vec2 textureCoordinate;" +
-                    "void main()" +
-                    "{"+
-                    "gl_Position = position;"+
-                    "textureCoordinate = inputTextureCoordinate;" +
-                    "}";
-
-    private final String fragmentShaderCode =
-            "#extension GL_OES_EGL_image_external : require\n"+
-                    "precision mediump float;" +
-                    "varying vec2 textureCoordinate;\n" +
-                    "uniform samplerExternalOES s_texture;\n" +
-                    "void main(void) {" +
-                    "  gl_FragColor = texture2D( s_texture, textureCoordinate );\n" +
-                    "}";
-
-    private FloatBuffer vertexBuffer, textureVerticesBuffer;
-    private ShortBuffer drawListBuffer;
-    private int mProgram;
-    private int mPositionHandle;
-    private int mColorHandle;
-    private int mTextureCoordHandle;
-
-
-    // number of coordinates per vertex in this array
-    static final int COORDS_PER_VERTEX = 2;
-    static float squareVertices[] = { // in counterclockwise order:
-            -1.0f, -1.0f,   // 0.left - mid
-            1.0f, -1.0f,   // 1. right - mid
-            -1.0f, 1.0f,   // 2. left - top
-            1.0f, 1.0f,   // 3. right - top
-    };
-
-    private short drawOrder[] =  {0, 2, 1, 1, 2, 3 }; // order to draw vertices
-    private short drawOrder2[] = {2, 0, 3, 3, 0, 1}; // order to draw vertices
-
-    static float textureVertices[] = {
-            0.0f, 1.0f,  // A. left-bottom
-            1.0f, 1.0f,  // B. right-bottom
-            0.0f, 0.0f,  // C. left-top
-            1.0f, 0.0f   // D. right-top
-    };
-
-    private final int vertexStride = COORDS_PER_VERTEX * 4; // 4 bytes per vertex
     private ByteBuffer indexBuffer;    // Buffer for index-array
     private int texture;
     private CardboardOverlayView mOverlayView;
@@ -97,6 +49,7 @@ public class MainActivity extends CardboardActivity implements CardboardView.Ste
     private SurfaceTexture surface;
     private float[] mView;
     private float[] mCamera;
+    private CameraEyeTransformer cameraPreviewTransformer;
 
     public void startCamera(int texture) {
         surface = new SurfaceTexture(texture);
@@ -119,59 +72,6 @@ public class MainActivity extends CardboardActivity implements CardboardView.Ste
         }
     }
 
-    static private int createTexture() {
-        int[] texture = new int[1];
-
-        GLES20.glGenTextures(1,texture, 0);
-        GLES20.glBindTexture(GL_TEXTURE_EXTERNAL_OES, texture[0]);
-        GLES20.glTexParameterf(GL_TEXTURE_EXTERNAL_OES,
-                GL10.GL_TEXTURE_MIN_FILTER,GL10.GL_LINEAR);
-        GLES20.glTexParameterf(GL_TEXTURE_EXTERNAL_OES,
-                GL10.GL_TEXTURE_MAG_FILTER, GL10.GL_LINEAR);
-        GLES20.glTexParameteri(GL_TEXTURE_EXTERNAL_OES,
-                GL10.GL_TEXTURE_WRAP_S, GL10.GL_CLAMP_TO_EDGE);
-        GLES20.glTexParameteri(GL_TEXTURE_EXTERNAL_OES,
-                GL10.GL_TEXTURE_WRAP_T, GL10.GL_CLAMP_TO_EDGE);
-
-        return texture[0];
-    }
-
-
-    private int loadGLShader(int type, String code) {
-        int shader = GLES20.glCreateShader(type);
-        GLES20.glShaderSource(shader, code);
-        GLES20.glCompileShader(shader);
-
-        // Get the compilation status.
-        final int[] compileStatus = new int[1];
-        GLES20.glGetShaderiv(shader, GLES20.GL_COMPILE_STATUS, compileStatus, 0);
-
-        // If the compilation failed, delete the shader.
-        if (compileStatus[0] == 0) {
-            Log.e(TAG, "Error compiling shader: " + GLES20.glGetShaderInfoLog(shader));
-            GLES20.glDeleteShader(shader);
-            shader = 0;
-        }
-
-        if (shader == 0) {
-            throw new RuntimeException("Error creating shader.");
-        }
-
-        return shader;
-    }
-
-    /**
-     * Checks if we've had an error inside of OpenGL ES, and if so what that error is.
-     * @param func
-     */
-    private static void checkGLError(String func) {
-        int error;
-        while ((error = GLES20.glGetError()) != GLES20.GL_NO_ERROR) {
-            Log.e(TAG, func + ": glError " + error);
-            throw new RuntimeException(func + ": glError " + error);
-        }
-    }
-
     /**
      * Sets the view to our CardboardView and initializes the transformation matrices we will use
      * to render our scene.
@@ -181,6 +81,7 @@ public class MainActivity extends CardboardActivity implements CardboardView.Ste
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
+        this.cameraPreviewTransformer = new CameraEyeTransformer();
         setContentView(R.layout.common_ui);
         cardboardView = (CardboardView) findViewById(R.id.cardboard_view);
         cardboardView.setRenderer(this);
@@ -227,37 +128,9 @@ public class MainActivity extends CardboardActivity implements CardboardView.Ste
     @Override
     public void onSurfaceCreated(EGLConfig config) {
         Log.i(TAG, "onSurfaceCreated");
-        GLES20.glClearColor(0.1f, 0.1f, 0.1f, 0.5f); // Dark background so text shows up well
+        cameraPreviewTransformer.createSurface();
 
-        ByteBuffer bb = ByteBuffer.allocateDirect(squareVertices.length * 4);
-        bb.order(ByteOrder.nativeOrder());
-        vertexBuffer = bb.asFloatBuffer();
-        vertexBuffer.put(squareVertices);
-        vertexBuffer.position(0);
-
-
-        ByteBuffer dlb = ByteBuffer.allocateDirect(drawOrder.length * 2);
-        dlb.order(ByteOrder.nativeOrder());
-        drawListBuffer = dlb.asShortBuffer();
-        drawListBuffer.put(drawOrder);
-        drawListBuffer.position(0);
-
-
-        ByteBuffer bb2 = ByteBuffer.allocateDirect(textureVertices.length * 4);
-        bb2.order(ByteOrder.nativeOrder());
-        textureVerticesBuffer = bb2.asFloatBuffer();
-        textureVerticesBuffer.put(textureVertices);
-        textureVerticesBuffer.position(0);
-
-        int vertexShader = loadGLShader(GLES20.GL_VERTEX_SHADER, vertexShaderCode);
-        int fragmentShader = loadGLShader(GLES20.GL_FRAGMENT_SHADER, fragmentShaderCode);
-
-        mProgram = GLES20.glCreateProgram();             // create empty OpenGL ES Program
-        GLES20.glAttachShader(mProgram, vertexShader);   // add the vertex shader to program
-        GLES20.glAttachShader(mProgram, fragmentShader); // add the fragment shader to program
-        GLES20.glLinkProgram(mProgram);
-
-        texture = createTexture();
+        texture = CameraEyeTransformer.createTexture();
         startCamera(texture);
     }
 
@@ -268,7 +141,7 @@ public class MainActivity extends CardboardActivity implements CardboardView.Ste
     @Override
     public void onNewFrame(HeadTransform headTransform) {
         float[] mtx = new float[16];
-        GLES20.glClear(GLES20.GL_COLOR_BUFFER_BIT | GLES20.GL_DEPTH_BUFFER_BIT);
+        cameraPreviewTransformer.clearGL();
         surface.updateTexImage();
         surface.getTransformMatrix(mtx);
     }
@@ -285,33 +158,7 @@ public class MainActivity extends CardboardActivity implements CardboardView.Ste
      */
     @Override
     public void onDrawEye(EyeTransform transform) {
-        GLES20.glClear(GLES20.GL_COLOR_BUFFER_BIT | GLES20.GL_DEPTH_BUFFER_BIT);
-
-        GLES20.glUseProgram(mProgram);
-
-        GLES20.glActiveTexture(GL_TEXTURE_EXTERNAL_OES);
-        GLES20.glBindTexture(GL_TEXTURE_EXTERNAL_OES, texture);
-
-        mPositionHandle = GLES20.glGetAttribLocation(mProgram, "position");
-        GLES20.glEnableVertexAttribArray(mPositionHandle);
-        GLES20.glVertexAttribPointer(mPositionHandle, COORDS_PER_VERTEX, GLES20.GL_FLOAT,
-                false,vertexStride, vertexBuffer);
-
-        mTextureCoordHandle = GLES20.glGetAttribLocation(mProgram, "inputTextureCoordinate");
-        GLES20.glEnableVertexAttribArray(mTextureCoordHandle);
-        GLES20.glVertexAttribPointer(mTextureCoordHandle, COORDS_PER_VERTEX, GLES20.GL_FLOAT,
-                false,vertexStride, textureVerticesBuffer);
-
-        mColorHandle = GLES20.glGetAttribLocation(mProgram, "s_texture");
-
-        GLES20.glDrawElements(GLES20.GL_TRIANGLES, drawOrder.length,
-                GLES20.GL_UNSIGNED_SHORT, drawListBuffer);
-
-
-        // Disable vertex array
-        GLES20.glDisableVertexAttribArray(mPositionHandle);
-        GLES20.glDisableVertexAttribArray(mTextureCoordHandle);
-
+        cameraPreviewTransformer.drawEye(texture);
         Matrix.multiplyMM(mView, 0, transform.getEyeView(), 0, mCamera, 0);
     }
